requirements.txt
# LangChain核心包
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2

# LangChain实验性包
langchain-experimental==0.0.50

# LLM提供商
openai>=1.3.0
anthropic>=0.7.0

# 数据库和向量存储
chromadb==0.4.18
faiss-cpu==1.7.4

# 工具包
wikipedia==1.4.0
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0

# 其他依赖
tiktoken==0.5.2
numpy==1.24.3
pandas==2.1.3
pillow==10.1.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1